IDEA的建议？

+++ raknet/src/main/java/com/nukkitx/network/VarInts.java
@@ -7,7 +7,7 @@ import lombok.experimental.UtilityClass;
 public class VarInts {
 
     public static void writeInt(ByteBuf buffer, int value) {
-        encode(buffer, ((value << 1) ^ (value >> 31)) & 0xFFFFFFFFL);
+        encode(buffer, (((long) value << 1) ^ (value >> 31)) & 0xFFFFFFFFL);
     }
 
     public static int readInt(ByteBuf buffer) {
